[Update] Minor webserver updates
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@3155 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/web/session/HttpSession.h
@@ -25,12 +25,12 @@ public:
 	String queryString;
 	String remoteUser;
 	String logMessage;
-	uint32 remoteIp;
+	uint64 remoteIp;
 	short remotePort;
 	short statusCode;
 	bool ssl;
 	Vector<HttpHeader*> headers;
-	uint32 timestamp;
+	uint64 timestamp;
 
     HttpSession(const struct mg_request_info *request_info);
 	virtual ~HttpSession();
@@ -64,7 +64,7 @@ public:
         return queryString;
     }
 
-    uint32 getRemoteIp() const
+    uint64 getRemoteIp() const
     {
         return remoteIp;
     }
